preprocessing-unstructured-data-for-llm-applications/requirements.txt
chromadb==1.5.1
langchain==1.2.10
langchain-community==0.4.1
langchain-openai==1.1.10
openai==2.24.0
tiktoken==0.12.0
#"unstructured[md,pdf,pptx]"
unstructured-client==0.42.10
unstructured==0.12.3
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==1.26.18
python-dotenv==1.0.1
panel==1.3.8
ipython==8.18.1
python-pptx==0.6.23
pdf2image==1.17.0
pdfminer.six==20251230
opencv-python==4.9.0.80
pikepdf==8.13.0
pypdf==6.2.0
matplotlib==3.7.2
pi-heif==1.3.0
